fix: export the correct methods
